requirements.txt
# MinerU Multimodal Contrastive Learning Data Factory
# Requirements for NTU EEE Cluster (4x A2000 GPUs)

# Core Dependencies
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
aiohttp>=3.9.0
Pillow>=10.0.0

# MinerU (PDF Parsing) - Install with: pip install mineru[all]
# Note: This is installed separately as it has complex dependencies
# mineru[all]>=2.0.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0

# Embeddings (optional, for hard negative mining)
sentence-transformers>=2.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Async Support
asyncio-throttle>=1.0.0

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
